Simplifty more of the query
